Add tabs in admin apps details
